refactor: 주문 테이블 정리 유효성 검사 이벤트 대체
