--- src/main/java/kitchenpos/tobe/common/domain/CustomEventPublisher.java
@@ -1,7 +1,0 @@
-package kitchenpos.tobe.common.domain;
-
-@FunctionalInterface
-public interface CustomEventPublisher<T> {
-
-    void publish(final T t);
-}
